Continued work on Generator component, updated regex for Externals and Subnets and moved features into proper folder structure
